undo refactor for interface
