+++ base/src/main/java/me/lambdaurora/mcpatcherpatcher/MCPatcherPatcherApplication.java
@@ -18,8 +18,6 @@
 package me.lambdaurora.mcpatcherpatcher;
 
 import me.lambdaurora.mcpatcherpatcher.image.BufferedImageProvider;
-import me.lambdaurora.mcpatcherpatcher.utils.CustomOutputStream;
-import me.lambdaurora.mcpatcherpatcher.utils.Dim2i;
 
 import javax.swing.*;
 import javax.swing.border.LineBorder;
@@ -27,6 +25,7 @@ import javax.swing.filechooser.FileNameExtensionFilter;
 import java.awt.*;
 import java.io.File;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.io.PrintStream;
 import java.util.Arrays;
 
@@ -225,6 +224,69 @@ public class MCPatcherPatcherApplication extends JFrame
                 patcherInterface.convert(inputFile, outputFile);
                 System.out.printf("Output File: %s%n", outputFile.getAbsolutePath());
             }
+        }
+    }
+
+    private class CustomOutputStream extends OutputStream
+    {
+        private final JTextArea textArea;
+
+        public CustomOutputStream(JTextArea textArea)
+        {
+            this.textArea = textArea;
+        }
+
+        @Override
+        public void write(int b)
+        {
+            textArea.append(String.valueOf((char) b));
+            textArea.setCaretPosition(textArea.getDocument().getLength());
+        }
+    }
+
+    private class Dim2i
+    {
+        private final int x;
+        private final int y;
+        private final int width;
+        private final int height;
+
+        public Dim2i(int x, int y, int width, int height)
+        {
+            this.x = x;
+            this.y = y;
+            this.width = width;
+            this.height = height;
+        }
+
+        public int getOriginX()
+        {
+            return this.x;
+        }
+
+        public int getOriginY()
+        {
+            return this.y;
+        }
+
+        public int getWidth()
+        {
+            return this.width;
+        }
+
+        public int getHeight()
+        {
+            return this.height;
+        }
+
+        public int getLimitX()
+        {
+            return this.x + this.width;
+        }
+
+        public int getLimitY()
+        {
+            return this.y + this.height;
         }
     }
 }
