--- base/src/main/java/me/lambdaurora/mcpatcherpatcher/utils/CustomOutputStream.java
@@ -1,38 +1,0 @@
-/*
- *  Copyright (c) 2020 LambdAurora <aurora42lambda@gmail.com>
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package me.lambdaurora.mcpatcherpatcher.utils;
-
-import javax.swing.*;
-import java.io.OutputStream;
-
-public class CustomOutputStream extends OutputStream
-{
-    private final JTextArea textArea;
-
-    public CustomOutputStream(JTextArea textArea)
-    {
-        this.textArea = textArea;
-    }
-
-    @Override
-    public void write(int b)
-    {
-        textArea.append(String.valueOf((char) b));
-        textArea.setCaretPosition(textArea.getDocument().getLength());
-    }
-}
