--- base/src/main/java/me/lambdaurora/mcpatcherpatcher/utils/Dim2i.java
@@ -1,64 +1,0 @@
-/*
- *  Copyright (c) 2020 LambdAurora <aurora42lambda@gmail.com>
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Lesser General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package me.lambdaurora.mcpatcherpatcher.utils;
-
-public class Dim2i
-{
-    private final int x;
-    private final int y;
-    private final int width;
-    private final int height;
-
-    public Dim2i(int x, int y, int width, int height)
-    {
-        this.x = x;
-        this.y = y;
-        this.width = width;
-        this.height = height;
-    }
-
-    public int getOriginX()
-    {
-        return this.x;
-    }
-
-    public int getOriginY()
-    {
-        return this.y;
-    }
-
-    public int getWidth()
-    {
-        return this.width;
-    }
-
-    public int getHeight()
-    {
-        return this.height;
-    }
-
-    public int getLimitX()
-    {
-        return this.x + this.width;
-    }
-
-    public int getLimitY()
-    {
-        return this.y + this.height;
-    }
-}
